+++ src/Symfony/Bridge/Twig/Tests/Extension/TranslationExtensionTest.php
@@ -51,14 +51,14 @@ class TranslationExtensionTest extends TestCase
     {
         $this->expectException('Twig\Error\SyntaxError');
         $this->expectExceptionMessage('Unexpected token. Twig was looking for the "with", "from", or "into" keyword in "index" at line 3.');
-        $output = $this->getTemplate("{% trans \n\nfoo %}{% endtrans %}")->render();
+        $this->getTemplate("{% trans \n\nfoo %}{% endtrans %}")->render();
     }
 
     public function testTransComplexBody()
     {
         $this->expectException('Twig\Error\SyntaxError');
         $this->expectExceptionMessage('A message inside a trans tag must be a simple text in "index" at line 2.');
-        $output = $this->getTemplate("{% trans %}\n{{ 1 + 2 }}{% endtrans %}")->render();
+        $this->getTemplate("{% trans %}\n{{ 1 + 2 }}{% endtrans %}")->render();
     }
 
     public function getTransTests()