+++ src/Symfony/Component/DependencyInjection/Dumper/PhpDumper.php
@@ -269,7 +269,7 @@ EOF;
                 foreach ($ids as $id) {
                     $c .= '    '.$this->doExport($id)." => true,\n";
                 }
-                $files['removed-ids.php'] = $c .= "];\n";
+                $files['removed-ids.php'] = $c."];\n";
             }
 
             if (!$this->inlineFactories) {